+++ tests/testthat/helper-tidyhydat.R
@@ -1,0 +1,6 @@
+skip_if_net_down <- function() {
+  if (has_internet()) {
+    return()
+  }
+  testthat::skip("no internet")
+}