+++ NEWS.md
@@ -3,6 +3,8 @@ tidyhydat 0.4.0
 
 ### IMPROVEMENTS
 * prov_terr_state_loc now accepts a "CA" value to specify only stations located in Canada (#112)
+* functions that access internet resources now fail with an informative error message (#116)
+* tests that require internet resources are skipped when internet is down 
 
 ### BUG FIXES
 * `pull_station_number` now only returns unique values (#109)